more coverage
also fix bug with element's required? check related to conditional
elements.  it was being set as not required when the conditional element
matched, when it should be setting it to not required only when the
conditional element doesn't match (If the conditional element does
match, it will default to the required flag on the element itself; that
way, we can have both optional or required elements that show up on
conditional of a previous element's anwer)