enforce page's all_element_ids order of ids
make use of this to make Fe::Page#complete? faster; also make views
required? checks faster by passing in the page so we don't have to
determine it in Fe::Element#previous_element (I could keep track of the
previous element in the view rendering and pass that in directly.. it
would be very slightly faster but not worth the code complexity.  With
all_element_ids col on pages, we can easily get the previous element if we
know the page anyways)