implement loading a page in the background
the page is validated so that the completion indicator is accurate (this
is necessary because when you active a page from a conditional question,
it's always marked as completed initially, because a non-required page
needs to be marked as completed or else the submit would fail.  So when
you enable a page as a result of answering a conditional question, if
that page has required questions, it needs to be evaluated so that the
incomplete indicator by the page name that CAP uses shows up.