handle conditional elements in a grid
makes use of all_elements_ids to make this easier

Came about because Canada has a series of questions in the emergency
info section.  Emergency info is a grid (to make it easier to clone)
and inside it is some questions where if your country is US or Canada is
asks you different health insurance questions.  Those I set up as
conditional questions but they weren't working  because they were in a
group