stop all the empty string answers
because the initial answers on the element were set using
responses(answer_sheet), that filters out empty strings.  So any update
where people hadn't chosen a dropdown value it was writing an empty
string.  You might want to check your database and see if there are a
lot of extra empty strings if you depend on pulling out answers
manually, like we are for apptrack in one-app/cap.  Using
display_response works because it already filters out empty strings