set initial answers to only the answer sheet's answers
solved by just setting the @answers instance var in question directly to
the answers instead of using ||=.  This was messing up some tests where
the same question instance was being used on different apps, so past
answers (for the wrong app) was still in @answers.  It seemed better to
just use @answers = than any other approach.

Added a sanity check that if set_response is called and it's not using the
object_name/attribute_name, ie. saving using the answers table, then
when you call save or save_file after that, it will check that the
answer sheets match