two minor fixes - js error fix, panel input field width too long
note that I could have used #panel in the css but since you can have
multiple panels open at once, using an id is a bit weird.  it probably
would work in most browsers but it's undefined behavior afaik.  q_form
class works and is a class instead of an id