allow hashes in the values
don't escape cgi escape the value in this case; this change enables
custom elements to pass in more detailed values.  This was motivated
from the panorama tool to make a question for which you can add or
delete any number of them.  This change enables an easy way to pass in
an arbitrary amount of answers