fix js error about options already defined
Fix some questions getting locked with a check that was too broad --
summer missions for example has an "apply" action that renders a form
and if a checkbox is on the first page that will always be locked
without this fix.

I think this was never noticed until now because almost all
questionnaires have an intro page, then future pages will get an ajax
get to the edit action and it will render correctly then.